requirements.txt
Flask==3.0.2
pypdf==4.1.0
google-generativeai==0.3.2
PyYAML==6.0.1 